pqc_crypto_package/requirements.txt

kyber-py>=1.0.1,<1.1.0
dilithium-py>=1.7,<2.0.0
pycryptodomex>=3.10.0,<4.0.0
aes256-ctr-drbg>=1.0.0,<2.0.0 

